// Front-end/src/app/Interfaces/category.ts
export interface Category {     // Exporta una interfaz llamada 'Category' que define la estructura de un objeto de categoría
    idCategory: number,         // Propiedad que representa el identificador de la categoría
    name: string                // Propiedad que representa el nombre de la categoría
}